server/requirements.txt
flask
mysql-connector-python==9.0.0
python-dotenv
flask-sqlalchemy==3.1.1
pyjwt
pymysql
python-kasa
asyncio
fastapi
uvicorn
flask-socketio